% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebsblacktheme.R
\name{ebs_black}
\alias{ebs_black}
\title{ebsblacktheme}
\usage{
ebs_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
create a plot that uses ggthemes and is dark in nature
}
\details{
ggtheme that has a dark background
}
\examples{
\dontrun{ ggplot + ebs_black()}
}
\seealso{
Other themes: 
\code{\link{ebs_white}()}
}
\concept{themes}
