% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebswhitetheme.R
\name{ebs_white}
\alias{ebs_white}
\title{ebswhitetheme}
\usage{
ebs_white(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
create a plot that uses ggthemes and is light  in nature
}
\details{
ggtheme that has a light background
}
\examples{
\dontrun{ ggplot + ebs_white()}
}
\seealso{
Other themes: 
\code{\link{ebs_black}()}
}
\concept{themes}
