% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\docType{data}
\name{ebsgreen}
\alias{ebsgreen}
\title{ebs green}
\format{An object of class \code{character} of length 1.}
\usage{
ebsgreen
}
\description{
ebs green
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_bar(aes(x = Sepal.Length), fill = ebsgreen) + ebs_black())
}
}
\keyword{datasets}
