% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\docType{data}
\name{ebsred}
\alias{ebsred}
\title{ebs red}
\format{An object of class \code{character} of length 1.}
\usage{
ebsred
}
\description{
ebs red
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_bar(aes(x = Sepal.Length), fill = #8B0000) + ebs_black())
}
}
\keyword{datasets}
