% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installdependents.R
\name{installIfAbsentAndLoad}
\alias{installIfAbsentAndLoad}
\title{instaldependents}
\usage{
installIfAbsentAndLoad(neededVector)
}
\description{
instal dependent package
}
\details{
dependent packages
}
\examples{
\donttest{
needed <- c('dplyr')
installIfAbsentAndLoad(needed)
}
}
