% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\docType{data}
\name{pal_ebs}
\alias{pal_ebs}
\title{colors}
\format{An object of class \code{character} of length 4.}
\usage{
pal_ebs
}
\description{
colors
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_point(aes(x = Sepal.Length, y = Sepal.Width, color= Species)) +scale_color_manual(values = pal_ebs) + ebs_black())
}
}
\keyword{datasets}
