% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\name{scale_color_continuous_ebs}
\alias{scale_color_continuous_ebs}
\title{continuous palette}
\usage{
scale_color_continuous_ebs()
}
\description{
continuous palette
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Sepal.Width)) +scale_color_continuous_ebs() + ebs_black())
}
}
