% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\name{scale_color_discrete_ebs}
\alias{scale_color_discrete_ebs}
\title{discrete palette}
\usage{
scale_color_discrete_ebs()
}
\description{
discrete palette
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +scale_color_discrete_ebs() + ebs_black())
}
}
