% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalecolor.R
\name{scale_fill_ebs}
\alias{scale_fill_ebs}
\title{colors}
\usage{
scale_fill_ebs()
}
\description{
colors
}
\details{
ebs colors
}
\examples{
\donttest{
(ggplot(iris) +geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +scale_fill_ebs() + ebs_black())
}
}
