% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_dark.R
\name{table_dark}
\alias{table_dark}
\title{table_dark}
\usage{
table_dark(data)
}
\arguments{
\item{data}{is a data frame}
}
\value{
The table formatted for the dark HTML theme
}
\description{
dark table for rmd html
}
\examples{
\dontrun{
data(iris, package = "datasets")
table_dark(iris)}
}
