% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_light.R
\name{table_light}
\alias{table_light}
\title{table_light}
\usage{
table_light(data)
}
\arguments{
\item{data}{Data frame used to make table and format in light theme}
}
\value{
The table formatted for the light HTML theme
}
\description{
create a table in RMarkdown for the trove light HTML theme
}
\examples{
\dontrun{
data(iris, package = "datasets")
table_light(iris)}
}
