% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pdf.R
\name{table_pdf}
\alias{table_pdf}
\title{table_pdf}
\usage{
table_pdf(data)
}
\arguments{
\item{data}{Data frame used to make table and format in pdf}
}
\value{
The table formatted for the pdf
}
\description{
create a table in RMarkdown for the trove pdf
}
\examples{
\dontrun{
data(iris, package = "datasets")
table_dark(iris)}
}
